feat|refactor: header reserved menu
