+++ src/partials/Header/Header.tsx
@@ -1,17 +1,48 @@
-import React, { useState } from 'react';
 import { graphql, Link, useStaticQuery } from 'gatsby';
+import React, { useState } from 'react';
 
-import { MenuMain } from './MenuMain';
-import { MenuReservedArea } from './MenuReservedArea';
-import { Socials } from '../Socials';
-import { Logo } from '../Logo';
 import { Hamburger } from '../Hamburger';
+import { Logo } from '../Logo';
+import { Socials } from '../Socials';
 
 import './Header.sass';
+import { useTranslation } from 'gatsby-plugin-react-i18next';
+import { Menu } from './Menu';
+import { useLocalizedQuery } from '../../hooks/useLocalizedQuery';
+
+enum MENU {
+  RESERVED_MENU = 'ReservedMenu',
+  MAIN_MENU = 'MainMenu',
+}
 
 export const Header = () => {
   const [mobileMenuOpen, setMobileMenuOpen] = useState<boolean>(false);
   const handleMobileMenu = () => setMobileMenuOpen(prev => !prev);
+
+  const query: Queries.MainNavigationQuery = useStaticQuery(graphql`
+    query MainNavigation {
+      allStrapiNavigation {
+        nodes {
+          ...MainNavigationItem
+        }
+      }
+    }
+  `);
+
+  console.debug(query);
+
+  const { localeNodes: menuNodes } = useLocalizedQuery<
+    Queries.MainNavigationItemFragment,
+    Queries.MainNavigationQuery
+  >({
+    query,
+    type: 'allStrapiNavigation',
+  });
+
+  const reservedMenu = menuNodes?.filter(
+    node => node?.key === MENU.RESERVED_MENU
+  );
+  const mainMenu = menuNodes?.filter(node => node?.key === MENU.MAIN_MENU);
 
   return (
     <header className={`header${mobileMenuOpen ? ' menu-is-open' : ''}`}>
@@ -28,7 +59,7 @@ export const Header = () => {
             </div>
 
             <div className="col-auto d-none d-lg-block">
-              <MenuReservedArea />
+              <Menu menu={reservedMenu} />
             </div>
           </div>
         </div>
@@ -38,7 +69,7 @@ export const Header = () => {
         <div className="container-fluid">
           <div className="row align-items-center justify-content-between">
             <div className="col-auto">
-              <MenuMain />
+              <Menu menu={mainMenu} />
             </div>
 
             <div className="col-auto d-none d-lg-block">
