+++ src/partials/Header/Menu.sass
@@ -231,7 +231,7 @@
           transform: translateY(-50%)
           transform-origin: center center
           right: .7rem
-          background-image: url(../../../images/ui/chevron-down-white.svg)
+          background-image: url(../../images/ui/chevron-down-white.svg)
           background-repeat: no-repeat
           background-size: 100%
           transform-origin: center center
@@ -331,7 +331,7 @@
 
         &.w-sub
           &::after
-            background-image: url(../../../images/ui/chevron-down-black.svg)
+            background-image: url(../../images/ui/chevron-down-black.svg)
 
   @media (max-width: $mobile-xs)
     padding: 4rem 2.4rem calc(#{$header-height-mobile} + 2rem)
