+++ src/partials/Header/Menu/index.ts
@@ -1,0 +1,1 @@
+export { Menu } from './Menu';
