+++ src/partials/Header/MenuItem/index.ts
@@ -1,0 +1,1 @@
+export { MenuItem } from './MenuItem';
