--- src/partials/Header/MenuMain/index.ts
@@ -1,1 +1,0 @@
-export { MenuMain } from './MenuMain';
