+++ src/partials/Header/MenuNavigation/index.ts
@@ -1,0 +1,1 @@
+export { MenuNavigation } from './MenuNavigation';
