--- src/partials/Header/MenuReservedArea/index.ts
@@ -1,1 +1,0 @@
-export { MenuReservedArea } from './MenuReservedArea';
